+++ Sources/Controllers/ALKConversationViewController+TableView.swift
@@ -10,6 +10,7 @@ import Foundation
 import UIKit
 import AVFoundation
 import Applozic
+import WebKit
 
 extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSource {
 
@@ -351,12 +352,30 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
                 }
                 return cell
             }
+
+        case .email:
+            if message.isMyMessage {
+                let cell: ALKMyEmailCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
+                cell.emailView.setWebViewDelegate(delegate: self, tag: indexPath.section)
+                cell.setBackgroundColor(UIColor.clear)
+                cell.update(viewModel: message)
+                cell.updateHeight(contentHeights[message.identifier])
+                return cell
+            } else {
+                let cell: ALKFriendEmailCell = tableView.dequeueReusableCell(forIndexPath: indexPath)
+                cell.emailView.setWebViewDelegate(delegate: self, tag: indexPath.section)
+                cell.setBackgroundColor(UIColor.clear)
+                cell.update(viewModel: message)
+                cell.updateHeight(contentHeights[message.identifier])
+                return cell
+            }
         }
     }
 
 
     public func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        return viewModel.heightForRow(indexPath: indexPath, cellFrame: self.view.frame)
+
+        return viewModel.heightForRow(indexPath: indexPath, cellFrame: self.view.frame,contentHeights: contentHeights)
     }
 
     public func tableView(_ tableView: UITableView, heightForHeaderInSection section: Int) -> CGFloat {
@@ -472,7 +491,7 @@ extension ALKConversationViewController: UITableViewDelegate, UITableViewDataSou
             contentOffsetDictionary[collectionView.tag] = horizontalOffset as AnyObject
         }
     }
-    
+
 }
 
 extension ALTopicDetail: ALKContextTitleDataType {
@@ -504,6 +523,7 @@ extension ALTopicDetail: ALKContextTitleDataType {
         }
         return "\(key): \(value)"
     }
+
 
 }
 
@@ -570,4 +590,45 @@ extension ALKConversationViewController: UICollectionViewDataSource,UICollection
 
     }
 
+}
+
+
+// MARK: - WKNavigationDelegate
+extension ALKConversationViewController:WKNavigationDelegate{
+
+    func updateEmailCell(_ cell: UITableViewCell, withHeight height: CGFloat) {
+        tableView.beginUpdates()
+        tableView.endUpdates()
+        UIView.animate(withDuration: 0.3) {
+            if let friendCell = cell as? ALKFriendEmailCell {
+                friendCell.updateHeight(height)
+            } else if let myCell = cell as? ALKMyEmailCell {
+                myCell.updateHeight(height)
+            }
+        }
+    }
+
+    public func webView(_ webView: WKWebView, didFinish navigation: WKNavigation!) {
+        guard
+            let message = viewModel.messageForRow(indexPath: IndexPath(row: 0, section: webView.tag)),
+            let cell = tableView.cellForRow(at: IndexPath(row: 0, section: webView.tag))
+            else {
+            return
+        }
+
+        //We already know the height of webview return from here
+        if let height = contentHeights[message.identifier] {
+            updateEmailCell(cell, withHeight: height)
+            return
+        }
+
+        if webView.scrollView.contentSize.height != 0 {
+            contentHeights[message.identifier] = webView.scrollView.contentSize.height
+        }
+        updateEmailCell(cell, withHeight: webView.scrollView.contentSize.height)
+    }
+
+    private func webView(_ webView: WKWebView, decidePolicyFor navigationAction: WKNavigationAction, decisionHandler: @escaping (WKNavigationActionPolicy) -> Void) {
+        decisionHandler(.cancel)
+    }
 }